Applied parts of Sergei Grechanik's kitty graphics protocol patch
 - http://st.suckless.org/patches/kitty-graphics-protocol/

The patch didn't apply without numerous conflicts, but cherry-picking
individual changes seemed to have worked. The underline parts were left
out as this fork already has different patches applied that implement
it.

There is one issue: the dummy characters that are rendered below the
image show up in random colors and the edges of the characters sometimes
bleed outside of the image border.